nas/requirements.txt
Flask
Flask-SQLAlchemy
flask-jwt-extended
psycopg2-binary
bootstrap-flask
sqlalchemy-mixins